[MOSIP-44662]: display Bio Extractors, Credential Type
