+++ pmp-ui-v2/src/pages/common/ApproveRejectPopup.js
@@ -1,24 +1,35 @@
-import React, { useState, useEffect } from 'react';
+import React, { useMemo, useState, useEffect } from 'react';
 import { useTranslation } from 'react-i18next';
-import { getPartnerManagerUrl, handleServiceErrors, createRequest, handleEscapeKey } from '../../utils/AppUtils';
+import { getPartnerManagerUrl, handleServiceErrors, createRequest, handleEscapeKey, getPartnerTypeDescription, isLangRTL } from '../../utils/AppUtils';
 import { HttpService } from '../../services/HttpService';
 import LoadingIcon from './LoadingIcon';
 import ErrorMessage from './ErrorMessage';
 import close_icon from '../../svg/close_icon.svg';
 import FocusTrap from 'focus-trap-react';
 import PropTypes from 'prop-types';
+import { getUserProfile } from '../../services/UserProfileService';
+import CredentialPartnerPolicyDetails from './CredentialPartnerPolicyDetails';
 
 function ApproveRejectPopup({ popupData, closePopUp, approveRejectResponse, title, subtitle, header, description }) {
     const { t } = useTranslation();
     const [errorCode, setErrorCode] = useState('');
     const [errorMsg, setErrorMsg] = useState('');
     const [dataLoaded, setDataLoaded] = useState(true);
+    const isLoginLanguageRTL = isLangRTL(getUserProfile().locale);
+
+    const isCredentialPolicyRequest = useMemo(() => {
+        const isPolicy = Boolean(popupData?.isPartnerPolicyRequest);
+        const partnerType = (popupData?.partnerType ?? '').toString().toUpperCase();
+        return isPolicy && partnerType === 'CREDENTIAL_PARTNER';
+    }, [popupData?.isPartnerPolicyRequest, popupData?.partnerType]);
+
+    const [policyDetailsLoading, setPolicyDetailsLoading] = useState(false);
 
     useEffect(() => {
+        const previousOverflow = document.body.style.overflow;
         document.body.style.overflow = "hidden";
-
         return () => {
-            document.body.style.overflow = "auto";
+            document.body.style.overflow = previousOverflow;
         };
     }, []);
 
@@ -109,41 +120,77 @@ function ApproveRejectPopup({ popupData, closePopUp, approveRejectResponse, titl
         innerDiv: "!flex !justify-between !items-center !rounded-xl !min-h-12 !p-3 !m-1 !-mb-2 w-full"
     }
 
+    const modalWidth = isCredentialPolicyRequest ? "md:w-[48rem] w-[95%]" : "md:w-[24rem] w-[55%]";
+
     return (
         <div className="fixed inset-0 flex items-center justify-center bg-black bg-opacity-35 z-50 font-inter cursor-default mx-1 break-normal">
             <FocusTrap focusTrapOptions={{ initialFocus: false, allowOutsideClick: true }}>
-                <div className="bg-white md:w-[24rem] w-[55%] mx-auto rounded-lg shadow-sm h-fit">
+                <div className={`bg-white mx-auto rounded-lg shadow-sm ${modalWidth} max-h-[90vh] flex flex-col`}>
                     {!dataLoaded ? (
                         <LoadingIcon styleSet={{ loadingDiv: '!py-[35%]' }} />
                     ) : (
                         <>
-                            <div className="relative">
+                            <div className="relative flex flex-col max-h-[90vh]">
                                 {errorMsg && (
                                     <ErrorMessage id='approve_reject_popup_error_msg' errorCode={errorCode} errorMessage={errorMsg} clickOnCancel={cancelErrorMsg} customStyle={customStyle} />
                                 )}
                                 <>
-                                    <div className="flex justify-between px-[1rem] my-4 w-full">
-                                        <div className="flex-col space-y-2 break-words w-[19.5rem]">
+                                    <div className="relative px-[1rem] pt-4 pb-2 w-full">
+                                        <div className="flex-col space-y-1 break-words pr-12 text-left">
                                             <p id='approve-reject_popup_title' className="text-sm font-bold">{title}</p>
                                             {subtitle && (
                                                 <p id='approve-reject_popup_sub_title' className="text-[#A5A5A5] text-xs">{subtitle}</p>
                                             )}
                                         </div>
-                                        <button id="approve_reject_popup_close_icon" onClick={closingPopUp} className="h-6 hover:cursor-pointer">
-                                            <img src={close_icon} alt="close" />
+                                        <button
+                                            id="approve_reject_popup_close_icon"
+                                            aria-label={t('commons.close')}
+                                            onClick={closingPopUp}
+                                            className="h-8 w-8 rounded-full bg-[#F2F4F8] flex items-center justify-center hover:cursor-pointer absolute top-3 right-4"
+                                        >
+                                            <img src={close_icon} alt="" className="h-4 w-4" />
                                         </button>
                                     </div>
                                     <hr className="h-px bg-gray-100 border-[0.02rem]" />
-                                    <div className="px-[1.5rem] py-3 text-center break-words">
-                                        <p id='approve-reject_popup_header' className="text-base font-semibold text-black">{header}</p>
-                                        <p id='approve-reject_popup_description' className="text-sm text-[#666666] py-3">{description}</p>
+                                    <div className="flex-1 overflow-y-auto">
+                                        <div className="px-[1.5rem] py-3 text-center break-words">
+                                            <p id='approve-reject_popup_header' className="text-base font-semibold text-black">{header}</p>
+                                            <p id='approve-reject_popup_description' className="text-sm text-[#666666] py-3">{description}</p>
+                                        </div>
+
+                                        {isCredentialPolicyRequest && (
+                                            <div className="px-[1.5rem] pb-3">
+                                                <CredentialPartnerPolicyDetails
+                                                    t={t}
+                                                    isLoginLanguageRTL={isLoginLanguageRTL}
+                                                    partnerId={popupData?.partnerId}
+                                                    policyId={popupData?.policyId}
+                                                    partnerTypeLabel={getPartnerTypeDescription(popupData?.partnerType, t) ?? popupData?.partnerType ?? '-'}
+                                                    enabled={isCredentialPolicyRequest}
+                                                    variant="popup"
+                                                    onLoadingChange={setPolicyDetailsLoading}
+                                                />
+                                            </div>
+                                        )}
                                     </div>
                                     <hr className="h-px bg-gray-100 border-[0.02rem]" />
                                     <div className="flex items-center justify-between space-x-3 p-[6%]">
-                                        <button id="reject_btn" onClick={() => handleStatusChange('rejected')} type="button" className="w-36 h-10 border-[#1447B2] border rounded-md text-tory-blue">
+                                        <button
+                                            id="reject_btn"
+                                            onClick={() => handleStatusChange('rejected')}
+                                            type="button"
+                                            disabled={policyDetailsLoading}
+                                            className={`w-36 h-10 border-[#1447B2] border rounded-md text-tory-blue ${policyDetailsLoading ? 'opacity-50 cursor-not-allowed' : ''}`}
+                                        >
                                             {t('approveRejectPopup.reject')}
                                         </button>
-                                        <button id="approve_btn" onClick={() => handleStatusChange('approved')} type="button" className="w-36 h-10 border-[#1447B2] border rounded-md bg-tory-blue text-white">
+                                        <button
+                                            id="approve_btn"
+                                            onClick={() => handleStatusChange('approved')}
+                                            type="button"
+                                            disabled={policyDetailsLoading}
+                                            className={`w-36 h-10 border-[#1447B2] border rounded-md bg-tory-blue text-white ${policyDetailsLoading ? 'opacity-50 cursor-not-allowed' : ''}`}
+                                        >
                                             {t('approveRejectPopup.approve')}
                                         </button>
                                     </div>
